+++ src/includedPlatforms.ts
@@ -2,14 +2,49 @@ import { PlatformsWiteSetting } from "./interfaces";
 
 export const includedPlatforms: PlatformsWiteSetting[] = [
   {
+    label: "Bing",
+    url: "https://www.bing.com/search?q=",
+    setting: "bing",
+  },
+  {
+    label: "crates.io",
+    url: "https://crates.io/search?q=",
+    setting: "cratesIo",
+  },
+  {
+    label: "CSS-Tricks",
+    url: "https://css-tricks.com/?s=",
+    setting: "cssTricks",
+  },
+  {
+    label: "dict.cc",
+    url: "https://www.dict.cc/?s=",
+    setting: "dictCc",
+  },
+  {
+    label: "Docs.rs",
+    url: "https://docs.rs/releases/search?query=",
+    setting: "docsRs",
+  },
+  {
     label: "Docker Hub",
     url: "https://hub.docker.com/search?q=",
     setting: "dockerHub",
   },
   {
+    label: "DuckDuckGo",
+    url: "https://duckduckgo.com/?q=",
+    setting: "duckDuckGo",
+  },
+  {
     label: "GitHub",
     url: "https://github.com/search?q=",
     setting: "gitHub",
+  },
+  {
+    label: "Google Translator",
+    url: "https://translate.google.com/?text=",
+    setting: "googleTranslator",
   },
   {
     label: "Google",
@@ -37,6 +72,11 @@ export const includedPlatforms: PlatformsWiteSetting[] = [
     setting: "nuGet",
   },
   {
+    label: "PyPI",
+    url: "https://pypi.org/search/?q=",
+    setting: "pyPI",
+  },
+  {
     label: "Reddit",
     url: "https://www.reddit.com/search?q=",
     setting: "reddit",
@@ -45,6 +85,11 @@ export const includedPlatforms: PlatformsWiteSetting[] = [
     label: "Stack Overflow",
     url: "https://stackoverflow.com/search?q=",
     setting: "stackOverflow",
+  },
+  {
+    label: "Startpage",
+    url: "https://www.startpage.com/do/dsearch?q=",
+    setting: "startpage",
   },
   {
     label: "Wikipedia",